** one big sdc svn merge (need heavy testing)  ** removed HubSettings class
** WARNING **
 this update will reset some of favhub settings

+++ trunk/client/HashBloom.cpp
@@ -66,7 +66,7 @@ size_t HashBloom::pos(const TTHValue& tth, size_t n) const {
 			x |= (1i64 << i);
 		}
 	}
-	return static_cast<size_t>(x % bloom.size());
+	return x % bloom.size();
 }
 
 void HashBloom::copy_to(ByteVector& v) const {
